removed content folder and MovieInfo component
