+++ src/features/MoviePage/Crew/index.js
@@ -1,7 +1,7 @@
 import { Title, PersonNavLink, Row } from './styled';
-import first from '../../images/person.png';
-import secend from '../../images/person2.png';
-import { PersonTile } from '../../../../common/PersonTile';
+import first from '../images/person.png';
+import secend from '../images/person2.png';
+import { PersonTile } from '../../../common/PersonTile';
 
 const Cast = () => (
     <>
